Ketaki: Style guide updated
