+++ Source/PushNotificationManager.swift
@@ -9,32 +9,23 @@
 import UIKit
 import UserNotifications
 
-enum NOTIFICATION:String{
-    case notificationRegister = "NotificationDidRegister"
-    case notificationReceive = "NotificationDidRecive"
-    case notificationFailedRegister = "NotificationDidFailedRegister"
-    func stringValue() -> String {
-        return self.rawValue
-    }
-}
 //MARK: Push Notification Manager
 class PushManager:NSObject
 {
     //:Class Variable Declaration:
-    var delegate:PushNotificationManagerDelegate?
     public var token:String? = nil
     public var isGranted:Bool = false
     static let shared = PushManager()
-    var objRegisterCompletion : registerCompletion!
-    var objReceiveCompletion : receiveCompletion!
+    var objRegisterCompletion : Register!
+    var objReceiveCompletion : Receive!
     
     private override init() { }
     //:Typlealias
-    typealias registerCompletion = (_ result: String?, _ error: Error? , _ isgranted: Bool?) -> Void
-    typealias receiveCompletion = ([AnyHashable : Any]) -> Void
+    typealias Register = (_ isgranted: Bool,_ token: String?, _ error: Error?) -> Void
+    typealias Receive  = ([AnyHashable : Any]) -> Void
     
     //:init overloading:
-    func setPushNotification(application: UIApplication,block:@escaping registerCompletion)
+    func set(_PushFor application: UIApplication,block:@escaping Register)
     {
         if #available(iOS 10, *)
         {
@@ -45,8 +36,7 @@ class PushManager:NSObject
                    //:Delgate call:
                     DispatchQueue.main.async
                     {
-                      //  self.delegate?.applicationdidDeniedPermission(application)
-                        block(nil,error,false)
+                        block(false,nil,error)
                     }
                     return
                 }
@@ -65,8 +55,7 @@ class PushManager:NSObject
                   //:Delgate call:
                     DispatchQueue.main.async
                     {
-                       // self.delegate?.applicationdidDeniedPermission(application)
-                        block(nil,nil,false)
+                        block(false,nil,nil)
                     }
                     return
                 }
@@ -81,7 +70,7 @@ class PushManager:NSObject
         }
        
     }
-    func subscribe(completion:@escaping receiveCompletion)
+    func subscribe(completion:@escaping Receive)
     {
         self.objReceiveCompletion = completion
     }
@@ -92,24 +81,41 @@ extension PushManager
     func ApplicationDidRegisterWithdeviceToken(_ application:UIApplication,deviceToken:Data)
     {
         token = deviceToken.map { String(format: "%02.2hhx", $0) }.joined()
-      //  let DeviceInfo:[String: String] = ["deviceToken": token!]
-         self.objRegisterCompletion(token!,nil,true)
-        //fire notifcation
-//        NotificationCenter.default.post(name: Notification.Name(NOTIFICATION.notificationRegister.stringValue()), object: nil, userInfo: DeviceInfo)
-//        self.delegate?.applicationRegisteredDeviceToken(application, deviceToken: deviceToken)
+         self.objRegisterCompletion(true,token!,nil)
     }
     func ApplicationdidFailedForRemoteNotification(_ application: UIApplication, error: Error)
     {
-        self.objRegisterCompletion(nil,error,false)
-//        NotificationCenter.default.post(name: Notification.Name(NOTIFICATION.notificationFailedRegister.stringValue()), object: nil, userInfo: nil)
-//        self.delegate?.applicationDidFailedRegisteredDeviceToken(application,error: error)
+        self.objRegisterCompletion(false,nil,error)
     }
     func ApplicationReceivedRemoteNotification(_ application: UIApplication?,data: [AnyHashable : Any])
     {
         self.objReceiveCompletion(data)
-//        NotificationCenter.default.post(name: Notification.Name(NOTIFICATION.notificationFailedRegister.stringValue()), object: nil, userInfo: data)
-//        self.delegate?.applicationReceiveRemoteNotification(application, userInfo: data)
     }
-    
 }
+//MARK: AppDelegate Extension
+extension AppDelegate
+{
+    //Degelate call
+    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data)
+    {
+        PushManager.shared.ApplicationDidRegisterWithdeviceToken(application, deviceToken: deviceToken)
+    }
+    func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error)
+    {
+        PushManager.shared.ApplicationdidFailedForRemoteNotification(application, error: error)
+    }
+    func application(_ application: UIApplication, didReceiveRemoteNotification data: [AnyHashable : Any])
+    {
+        PushManager.shared.ApplicationReceivedRemoteNotification(application,data: data)
+    }
+    func userNotificationCenter(_ center: UNUserNotificationCenter,  willPresent notification: UNNotification, withCompletionHandler   completionHandler: @escaping (_ options:   UNNotificationPresentationOptions) -> Void)
+    {
+        PushManager.shared.ApplicationReceivedRemoteNotification(nil,data: notification.request.content.userInfo)
+    }
+    func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void)
+    {
+        PushManager.shared.ApplicationReceivedRemoteNotification(nil,data: response.notification.request.content.userInfo)
+    }
+}
+
